Add files via upload
This is a Hotel Booking Web Application that allows users to browse available hotels, search by location, and book rooms seamlessly. Built with HTML, CSS, and JavaScript, it features a user-friendly interface and a responsive design for both desktop and mobile devices. The project highlights my ability to create interactive and dynamic web applications while focusing on a smooth user experience.